icsspulse-light/requirements.txt
pymodbus==3.11.2
opcua==0.98.13
flask==3.1.2
ansi2html==1.9.2
dotenv==0.9.9
python-snap7==2.1.0
paho-mqtt==2.1.0